added new changes to fix HW
added new changes to fix HW
