+++ .gitignore
@@ -14,3 +14,4 @@ gradle*
 # gradle/wrapper/gradle-wrapper.properties
 
 /out/
+src/test/java/Homework16.java
