--- src/test/java/Homework16.java
@@ -1,34 +1,0 @@
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import java.time.Duration;
-
-public class Homework16 extends LoginTests {
-@Test
-public void registrationNavigation() {
-
-    ChromeOptions options = new ChromeOptions();
-    options.addArguments("--remote-allow-origins=*");
-
-    WebDriver driver = new ChromeDriver(options);
-    driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-
-    String url = "https://qa.koel.app/";
-    driver.get(url);
-
-    WebElement submit = driver.findElement(By.cssSelector("href=registration"));
-    submit.click();
-
-    Assert.assertEquals(driver.getCurrentUrl(), "https://qa.koel.app/registration");
-
-    driver.quit();
-}
-}
-
-// Tried running this test and it failed so not sure what I did wrong above (if anything). Please advise! Thanks
-
